+++ src/main/java/duke.txt
@@ -1,2 +1,1 @@
-D | 0 | hello  | Dec 12 2021
-D | 0 | hi | Nov 11 2022
+D | 0 | return book  | Dec 12 2022
