+++ text-ui-test/EXPECTED.TXT
@@ -1,26 +1,27 @@
+Loading error...
 Hello! I'm SmartBot
 What can I do for you?
 Got it. I've added this task:
  [T][ ] borrow book
 Now you have 1 tasks in the list.
 Got it. I've added this task:
- [D][ ] return book  (by: Sunday)
+ [D][ ] return book  (by: Dec 12 2021)
 Now you have 2 tasks in the list.
 Got it. I've added this task:
  [E][ ] project meeting  (at: Mon 2-4pm)
 Now you have 3 tasks in the list.
 1. [T][ ] borrow book
-2. [D][ ] return book  (by: Sunday)
+2. [D][ ] return book  (by: Dec 12 2021)
 3. [E][ ] project meeting  (at: Mon 2-4pm)
 Nice! I've marked this task as done:
   [X] borrow book
 OK, I've marked this task as not done yet:
-  [ ] return book 
+  [ ] return book
 1. [T][X] borrow book
-2. [D][ ] return book  (by: Sunday)
+2. [D][ ] return book  (by: Dec 12 2021)
 3. [E][ ] project meeting  (at: Mon 2-4pm)
-Noted. I've removed this task: 
- [D][ ] return book  (by: Sunday)
+Noted. I've removed this task:
+ [D][ ] return book  (by: Dec 12 2021)
 Now you have 2 tasks in the list.
 1. [T][X] borrow book
 2. [E][ ] project meeting  (at: Mon 2-4pm)
